requirements.txt
# Polymarket Maker Bot 依赖包

# 核心依赖
aiohttp>=3.9.0
asyncio>=3.4.3
python-dotenv>=1.0.0

# WebSocket
websockets>=12.0
aiohttp-sse-client>=0.2.1

# 加密和签名
ecdsa>=0.18.0
eth-account>=0.9.0
web3>=6.0.0
pycryptodome>=3.19.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# 配置和日志
pyyaml>=6.0
structlog>=23.1.0
colorlog>=6.7.0

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.12.0

# API客户端
requests>=2.31.0
httpx>=0.25.0

# 时间处理
pytz>=2023.3
arrow>=1.2.3

# 监控和指标
prometheus-client>=0.17.0
psutil>=5.9.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# 开发工具
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
flake8>=6.0.0

# 其他工具
tenacity>=8.2.0  # 重试库
cachetools>=5.3.0
ujson>=5.8.0  # 快速JSON解析
orjson>=3.9.0